requirements.txt
transformers
sentencepiece
tqdm
loguru
pyYAML==5.3.1